raspberry-pi/requirements.txt
async-timeout==4.0.2
colorzero==2.0
Deprecated==1.2.13
gpiozero==1.6.2
packaging==21.3
pyparsing==3.0.9
raspberrypi-tm1637==1.3.6
redis==4.3.4
RPi.GPIO==0.7.1
wiringpi==2.60.1
wrapt==1.14.1